apps/python-grpc-server/pyproject.toml
[project]
name = "python-grpc-server"
version = "0.1.0"
description = "Simple gRPC server using betterproto"
requires-python = ">=3.8"
dependencies = [
    "grpclib==0.4.4",
    "types-protobuf>=4.24",
    "protobuf>=4.25.3",
    "grpcio>=1.56.2",
    "grpcio-tools>=1.56.2",
    "grpc-stubs>=1.53.0.2",
    "python-grpc-bindings",
]

[tool.uv]
dev-dependencies = []

[tool.uv.sources]
python-gRPC-bindings = { path = "../../packages/proto/generated/python", editable = true }
